requirements.txt
requests
pandas
mistralai
python-decouple
